+++ cli.js
@@ -2,4 +2,5 @@
 
 require("./lib/cjs/cli.js")
 	.cli()
+	// eslint-disable-next-line no-console
 	.catch(console.log);