+++ src/cli/analyzer-cli-config.ts
@@ -1,16 +1,28 @@
-import { AnalyzerConfig } from "../analyze/types/analyzer-config";
+import * as tsModule from "typescript";
+import { ComponentFeature } from "../analyze/types/features/component-feature";
 import { VisibilityKind } from "../analyze/types/visibility-kind";
 import { TransformerKind } from "../transformers/transformer-kind";
 
 export interface AnalyzerCliConfig {
+	glob?: string[];
+
+	dry?: boolean;
 	debug?: boolean;
 	outFile?: string;
+	outFiles?: string;
 	outDir?: string;
+
 	format?: TransformerKind;
-	discoverLibraryFiles?: boolean;
 	visibility?: VisibilityKind;
+
+	analyze?: {
+		features?: ComponentFeature[];
+		discoverLibraryFiles?: boolean;
+	};
+
 	markdown?: {
 		titleLevel?: number;
 	};
-	analyze?: AnalyzerConfig;
+
+	ts?: typeof tsModule;
 }