+++ src/cli/index.ts
@@ -1,3 +1,3 @@
 export * from "./cli";
 export * from "./analyzer-cli-config";
-export * from "./cli-command/analyze/analyze-cli-command";
+export * from "./analyze/analyze-cli-command";