+++ src/cli/util/command-error.ts
@@ -1,0 +1,1 @@
+export class CommandError extends Error {}